The test that justified the decision did not test the decision
Adversarial review of #2122. Verified every finding before acting; the first
one is the serious one and it was mine.

a_session_finding_is_covered_by_the_run_root was decorative. It built an event
by hand and called compute_content_hash and compute_run_root directly, so it
passed with the kind replaced by 'assay.totally.unrelated.kind' and passed with
Payload::SessionFinding deleted from the crate. It tested that a hash function
depends on its input, which crypto::id::tests::test_content_hash_changes_with_payload
already tested. The PR body called it 'the one that matters'.

Replaced with tests/session_finding_bundle_roundtrip.rs, which writes a bundle
through BundleWriter, verifies it with verify_bundle_with_limits, then edits the
finding's bytes inside events.ndjson and requires an Integrity failure. Bitten:
removing the finding payload from the bundle fails it.

The integrity argument reached the right answer by the wrong mechanism. A sibling
file sitting outside run_root proves nothing, because events.ndjson is not in
run_root either -- its bytes are covered by m.files[..].sha256, which a sibling
file would share -- and since ADR-044 the attestation subject is the digest of
the whole archive, which covers a sibling file outright. What actually decides it
is that verify_bundle never iterates m.files for completeness, so a manifest
listing an absent file verifies clean. Now stated, with the archive-digest
counter-argument, and the verifier gap named as a separate defect.

The OCSF attribute is is_alert (ocsf/ocsf-schema#1178, merged 2024-09-27), not
is_detection, which appears only in the issue's contested proposal. Issue #1177
is still open, so the past tense was wrong too.

The singleton-predicate rule is in-toto's, restated in OpenVEX's embedding guide.
The constraint binds us either way -- our seal is an in-toto predicate -- but the
authority was borrowed from the wrong document.

'Every consumer of ours reads through that view' was false: there are no
production consumers of Payload at all. The urgency claim is now stated at its
true size.

TraceExtent had no label(), no Serialize, no rendering anywhere, so 'complete'
and 'partial' were invented by the payload -- worse than duplicated, because
there was no source to drift from. TraceExtent::label() added, and
tests/session_finding_vocabulary_parity.rs pins both it and RuleOutcome::label()
against their definitions. assay-core cannot be called from here without a cycle,
so the parity test is the fallback CLAUDE.md sanctions and claim_gate_parity.rs
already uses in this crate.

spanned is u64, not usize: RuleEvaluation holds in-memory indices, a wire format
re-read on other machines must not be platform-width.

Also fixed: the Consequences bullet still claimed Unknown absorbed this class,
contradicting Decision 1 two sections above; #2123 and #2124 are now cited by
number; the CFD quote had its qualifier restored.

+++ crates/assay-evidence/src/types.rs
@@ -370,15 +370,29 @@ pub struct PayloadSessionFinding {
     pub kind: String,
     /// `held`, `violated` or `not_exercised`. Three values rather than a boolean, because a rule
     /// that never got to decide is not a rule that passed.
+    ///
+    /// The strings are `assay_core::sequence_eval::RuleOutcome::label()`, which is documented there
+    /// as an interface rather than a `Debug` view. This crate cannot call it -- `assay-core`
+    /// depends on `assay-evidence` through `assay-adapter-api`, so the edge would cycle -- so
+    /// `tests/session_finding_vocabulary_parity.rs` is the sanctioned fallback, the same shape
+    /// `tests/claim_gate_parity.rs` already uses for the claim gate.
+    ///
+    /// The type is `String` rather than an enum here on purpose: an evidence reader must be able to
+    /// deserialise a record written by a newer producer that added a fourth outcome, rather than
+    /// failing on it. The parity test is what keeps the three known spellings honest.
     pub outcome: String,
     /// The call indices this finding ranges over.
+    ///
+    /// `u64` rather than the producer's `usize`: those are in-memory indices, and this is a wire
+    /// format re-read by third parties on other machines, where platform width is not a property
+    /// worth inheriting.
     ///
     /// Meaningful relative to the call sequence the finding was computed over, and stated here
     /// rather than implied: these are positions, not content addresses, so they identify calls
     /// only alongside the run this event belongs to. Binding a span to event content hashes would
     /// make the record self-contained and is deliberately not done yet — the producer does not
     /// know those hashes at evaluation time, and inventing the link would outrun the layer.
-    pub spanned: Vec<usize>,
+    pub spanned: Vec<u64>,
     /// Whether the trace this was evaluated against was finished (`complete`) or may still grow
     /// (`partial`).
     ///
@@ -386,6 +400,10 @@ pub struct PayloadSessionFinding {
     /// on a finished run and undecided on a live one, and #2112 established that the difference is
     /// invisible in the rules and the trace — it is only in who is asking. A consumer that cannot
     /// see this field would have to guess.
+    ///
+    /// The strings are `assay_core::sequence_eval::TraceExtent::label()`, added with this ADR
+    /// because the enum had no rendering before it, and pinned by the same parity test as
+    /// `outcome`.
     pub extent: String,
     /// Why, in the producer's words. Present for `violated` and `not_exercised`; a held rule needs
     /// no prose, and inventing one would invite readers to parse it.
